requirements.txt
groq
requests
sqlalchemy
pandas
python-dotenv
newspaper3k
lxml[html_clean]